TINY-6505: Move sub components to card
