+++ modules/bridge/src/main/ts/ephox/bridge/components/menu/card/Description.ts
